requirements.txt
mnemonic
eth-keys 
eciespy